Backend/requirements.txt
google-genai
websockets>=11.0.0
pydub>=0.25.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.11.0